ran check-style & fixed the rest of the easy errors in the touched files. Now at 4389 errors
