+++ components/webrtc_view_bulma/TextChat.jsx
@@ -1,14 +1,18 @@
-import React from "react";
-import {connect, compose} from 'react-redux';
+// Copyright (c) 2018-present Riff Learning, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
+
+import React from 'react';
+import {connect} from 'react-redux';
 import {Widget, addResponseMessage, addUserMessage, dropMessages} from 'react-chat-widget';
 import styled from 'styled-components';
 import lifecycle from 'react-pure-lifecycle';
 import _ from 'underscore';
-import 'react-chat-widget/lib/styles.css';
-import {withReducer} from 'recompose';
 import {getCurrentUser} from 'mattermost-redux/selectors/entities/users';
-import { logger } from '../../utils/riff';
+
 import {sendTextChatMsg, setTextChatBadge} from '../../actions/webrtc_actions';
+import {logger} from '../../utils/riff';
+
+import 'react-chat-widget/lib/styles.css';
 
 const RiffChat = styled.div`
 .rcw-conversation-container > .rcw-header {
@@ -42,18 +46,18 @@ const mapStateToProps = (state) => ({
 const mapDispatchToProps = (dispatch) => ({
     dispatch,
     removeBadge: () => {
-        logger.debug("removing badge...");
+        logger.debug('removing badge...');
         dispatch(setTextChatBadge(false));
-    }
+    },
 });
 
 const componentDidMount = (props) => {
     dropMessages();
     _.each(props.messages, (m) => {
-        if (m.participant == props.uid) {
+        if (m.participant === props.uid) {
             addUserMessage(m.message);
         } else {
-            addResponseMessage("**" + m.name + "**" + ": " + m.message);
+            addResponseMessage(`**${m.name}**: ${m.message}`);
         }
     });
 };
@@ -63,27 +67,25 @@ const componentDidUpdate = (props, prevProps) => {
 
     function arrayDiff(a, b) {
         return [
-            ...a.filter(x => !b.includes(x)),
-            ...b.filter(x => !a.includes(x))
+            ...a.filter((x) => !b.includes(x)),
+            ...b.filter((x) => !a.includes(x)),
         ];
     }
-    if (props.messages != prevProps.messages && props.messages.length > 0) {
+    if (props.messages !== prevProps.messages && props.messages.length > 0) {
         let newMessages = arrayDiff(props.messages, prevProps.messages);
-        const numNewMessages = newMessages.length;
-        newMessages = _.filter(newMessages, (m) => { return m.participant != props.uid; });
+        newMessages = _.filter(newMessages, (m) => m.participant !== props.uid);
         logger.debug('new messages:', newMessages);
         _.each(newMessages, (m) => {
-            addResponseMessage('**' + m.name + '**' + ': ' + m.message);
+            addResponseMessage(`**${m.name}**: ${m.message}`);
             props.dispatch(setTextChatBadge(true));
         });
     }
 
-    if (props.messages.length == 0) {
+    if (props.messages.length === 0) {
         logger.debug('props messages EMPTY, clearing messages on chat component.');
         dropMessages();
     }
 };
-
 
 const mapMergeProps = (stateProps, dispatchProps, ownProps) => ({
     ...stateProps,
@@ -93,21 +95,23 @@ const mapMergeProps = (stateProps, dispatchProps, ownProps) => ({
         dispatchProps.dispatch(sendTextChatMsg(event,
                                                stateProps.uid,
                                                stateProps.riff.meetingId));
-    }
+    },
 });
 
 const methods = {
     componentDidMount,
-    componentDidUpdate
+    componentDidUpdate,
 };
 
 const ChatView = (props) => (
-    <RiffChat onClick={event => props.removeBadge()}>
-      <Widget handleNewUserMessage={props.handleNewUserMessage}
-              onClick={event => props.removeBadge()}
-        title=''
-        subtitle=''
-        badge={props.badge}/>
+    <RiffChat onClick={() => props.removeBadge()}>
+        <Widget
+            handleNewUserMessage={props.handleNewUserMessage}
+            onClick={() => props.removeBadge()}
+            title=''
+            subtitle=''
+            badge={props.badge}
+        />
     </RiffChat>
 );
